reorganize debug commands
